medical_services_chatbot/requirements.txt
streamlit>=1.24.0
openai>=1.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
azure-openai>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
seaborn>=0.12.0
matplotlib>=3.7.0
tqdm>=4.65.0
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
fastapi>=0.103.0
uvicorn>=0.23.2
httpx>=0.24.1
requests>=2.31.0
pydantic>=2.4.0
prometheus-client>=0.17.0
python-json-logger>=2.0.7
structlog>=23.1.0
psutil>=5.9.0